requirements.txt
# requirements.txt

groq
python-dotenv
requestsstreamlit
tiktoken